api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "little-lairs-api"
version = "2.0.0"
description = "Little Lairs Citation Forensics Benchmark API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "celery[redis]==5.4.0",
    "redis==5.2.1",
    "httpx==0.28.1",
    "tldextract==5.1.3",
    "beautifulsoup4==4.12.3",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-mock==3.14.0",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
